+++ packages/dartway_flutter/lib/src/dw/dw.dart
@@ -7,7 +7,6 @@ import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import 'services/dw_shared_preferences.dart';
-import 'telegram_app/telegram_app.dart';
 
 part 'parts/dw_navigation.dart';
 part 'parts/dw_notifications.dart';