+++ packages/dartway_flutter/lib/src/dw/parts/dw_services.dart
@@ -3,30 +3,12 @@ part of '../dw.dart';
 class _DwServices {
   _DwServices._();
 
-  // static final _DwServices _instance = _DwServices._();
-
-  // static _DwServices get i => _instance;
-
   late DwSharedPreferences _prefs;
   DwSharedPreferences get sharedPreferences => _prefs;
 
-  late final TelegramApp _telegramWebApp;
-  TelegramApp get telegramWebApp => _telegramWebApp;
-
-  // DotEnv get dotEnv => dotenv;
-
-  _init({required DwConfig config}) async {
+  Future<void> _init({required DwConfig config}) async {
     if (config.useSharedPreferences) {
       _prefs = DwSharedPreferences(await SharedPreferences.getInstance());
     }
-
-    if (config.telegramWebAppConfig != null) {
-      _telegramWebApp = const TelegramAppImpl();
-      await _telegramWebApp.init(config.telegramWebAppConfig!);
-    }
-
-    // if (config.flutterDotEnvFile != null) {
-    //   await dotenv.load(fileName: config.flutterDotEnvFile!);
-    // }
   }
 }